Correct stack overflow when init with bm driver device (Orbit reader 20)
There is a crash happens in create KeyBinding array: "Cause: stack pointer is in a non-existent map; likely due to stack overflow." while trying to connect with Orbit reader 20 in 64 bit CPU architecture which the size variable in this case is more than 15,000. It seems static creation is not enough for this case so we propose malloc solution and the stack overflow not exists anymore.